noItems view

+++ TodoList/Views/ListView.swift
@@ -15,22 +15,33 @@ struct ListView: View {
     
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-               
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item  )
-                             
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                Text("No items")
+            } else {
+                
+                List{
+                    ForEach(listViewModel.items) { item in
+                       
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item  )
+                                     
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                   
+                }
+                .listStyle(PlainListStyle() )
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-           
+            
+            
         }
-        .listStyle(PlainListStyle() )
+        
+        
         .navigationTitle("To Do List")
         .navigationBarItems(
             leading: EditButton(),
